+++ packages/integration-test/index.test.ts
@@ -1,17 +1,29 @@
-import stubImplementation from '@onting/stub/implementation';
+import createStubImplementation from '@onting/stub/host.js';
 import { scenario } from '@testduet/given-when-then';
 import { expect } from 'expect';
 import * as NodeTest from 'node:test';
 
 scenario(
-  'stubImplementation',
+  'createStubImplementation',
   bdd => {
     bdd
-      .given('a stub implementation', () =>
-        stubImplementation.implement(
+      .given('a stub implementation', async () =>
+        (
+          await createStubImplementation(() => ({
+            async getNext() {
+              return undefined;
+            },
+            async setCurrent() {}
+          }))
+        ).implement({
+          browsingContext: {
+            getTopLevelContexts() {
+              return [{ url: 'https://example.com' }];
+            }
+          },
+          webDriver: {}
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          {} as any
-        )
+        } as any)
       )
       .when('getTimestamp is called', precondition => precondition.getTimestamp())
       .then('should return a string starting with "Hello, World!"', (_, result) =>